Refactored adapters out of STH
