+++ bdd/package.json
@@ -12,7 +12,7 @@
     "find-package-json": "^1.2.0",
     "freeport": "^1.0.5",
     "n-readlines": "^1.0.1",
-    "scramjet": "^4.36.9"
+    "scramjet": "^4.37.0"
   },
   "devDependencies": {
     "@cucumber/cucumber": "^7.3.2",
