+++ packages/adapter-docker/.nycrc.json
@@ -1,0 +1,13 @@
+{
+    "all": true,
+    "include": [
+      "dist/**/*"
+    ],
+    "exclude": [
+        "**/*.spec.ts"
+    ],
+    "reporter": [
+        "lcovonly",
+        "text"
+    ]
+}
