+++ packages/adapter-docker/src/docker-networking.ts
