+++ packages/adapter-docker/src/dockerode-docker-helper.ts
