+++ packages/adapter-process/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@scramjet/load-check",
+  "name": "@scramjet/adapter-process",
   "version": "0.37.0",
-  "description": "This package is part of Scramjet Transform Hub. The package provides health info for STH.",
+  "description": "This package is part of Scramjet Transform Hub. This module holds the process adapters utilized by Scramjet Transform Hub",
   "main": "./src/index.ts",
   "scripts": {
     "start": "ts-node ./src/index",
@@ -9,21 +9,27 @@
     "build:docs": "typedoc",
     "clean": "rm -rf ./dist .bic_cache",
     "cloc": "cloc src --fullpath --include-lang TypeScript --not-match-d \"(node_modules|test|dist|bdd)\" --by-percent cm",
-    "test": "nyc ava"
+    "test": "nyc ava",
+    "prepack": "node ../../scripts/publish.js"
   },
   "author": "Scramjet <open-source@scramjet.org>",
   "license": "AGPL-3.0",
   "dependencies": {
+    "@scramjet/model": "^0.37.0",
     "@scramjet/obj-logger": "^0.37.0",
+    "@scramjet/pre-runner": "^0.37.0",
+    "@scramjet/python-runner": "^0.37.0",
+    "@scramjet/runner": "^0.37.0",
+    "@scramjet/sth-config": "^0.37.0",
+    "@scramjet/symbols": "^0.37.0",
     "@scramjet/utility": "^0.37.0",
-    "scramjet": "^4.37.0",
-    "systeminformation": "^5.21.7",
-    "uuid": "^8.3.2"
+    "@scramjet/adapters": "0.37.0",
+    "scramjet": "^4.37.0"
   },
   "devDependencies": {
     "@scramjet/types": "^0.37.0",
     "@types/node": "15.12.5",
-    "@types/uuid": "^8.3.4",
+    "@types/request": "2.48.8",
     "ava": "^3.15.0",
     "ts-node": "^10.9.1",
     "typedoc": "0.23.17",
