+++ packages/adapter-process/src/index.ts
@@ -1,0 +1,6 @@
+export { ProcessSequenceAdapter as SequenceAdapter } from "./process-sequence-adapter";
+export { ProcessInstanceAdapter as InstanceAdapter } from "./process-instance-adapter";
+
+export function initialize() {
+    return true;
+}
