+++ packages/adapter-process/src/readme.mtpl
@@ -1,0 +1,18 @@
+# Scramjet Transform Hub Adapters
+
+This module holds two types of adapters utilized by Scramjet Transform Hub: Instance Adapter and Sequence Adapter. These Adapters allows for running the Sequence identification and Instance execution in two basic modes: as a non containerized standalone processes or in a process container.
+
+The adapter provides two main exports:
+
+* [processSequenceAdapter](./src/process-sequence-adapter.ts) - An adapter for preparing Sequence to be run in process container.
+* [processInstanceAdapter](./src/process-instance-adapter.ts) - An adapter for running Instance by Runner executed in process container.
+
+## Docs
+
+>!docs adapters-process/modules.md &
+
+>@sth
+>@use-cases
+>@links
+>@license-agpl
+>@contrib
