+++ packages/adapters/package.json
@@ -15,29 +15,19 @@
   "author": "Scramjet <open-source@scramjet.org>",
   "license": "AGPL-3.0",
   "dependencies": {
-    "@kubernetes/client-node": "^0.17.1",
+    "@scramjet/adapter-docker": "^0.37.0",
+    "@scramjet/adapter-kubernetes": "^0.37.0",
+    "@scramjet/adapter-process": "^0.37.0",
     "@scramjet/model": "^0.37.0",
     "@scramjet/obj-logger": "^0.37.0",
-    "@scramjet/pre-runner": "^0.37.0",
-    "@scramjet/python-runner": "^0.37.0",
-    "@scramjet/runner": "^0.37.0",
     "@scramjet/sth-config": "^0.37.0",
     "@scramjet/symbols": "^0.37.0",
     "@scramjet/utility": "^0.37.0",
-    "dockerode": "^3.3.4",
-    "scramjet": "^4.36.9",
-    "shell-escape": "^0.2.0",
-    "systeminformation": "^5.21.7",
-    "ts.data.json": "^2.2.0"
+    "scramjet": "^4.37.0"
   },
   "devDependencies": {
     "@scramjet/types": "^0.37.0",
-    "@types/dockerode": "<=3.3.3",
-    "@types/js-yaml": "4.0.5",
     "@types/node": "15.12.5",
-    "@types/request": "2.48.8",
-    "@types/shell-escape": "^0.2.1",
-    "@types/ws": "8.5.3",
     "ava": "^3.15.0",
     "ts-node": "^10.9.1",
     "typedoc": "0.23.17",
