+++ packages/cli/package.json
@@ -30,7 +30,7 @@
     "commander-completion": "^1.0.1",
     "find-package-json": "^1.2.0",
     "minimatch": "^3.1.2",
-    "scramjet": "^4.36.9",
+    "scramjet": "^4.37.0",
     "tar": "^6.1.11",
     "validator": "^13.7.0"
   },
