+++ packages/host/package.json
@@ -36,7 +36,7 @@
     "minimist": "^1.2.6",
     "pico-s3": "^2.0.0",
     "rereadable-stream": "^1.4.14",
-    "scramjet": "^4.36.9",
+    "scramjet": "^4.37.0",
     "systeminformation": "^5.21.7"
   },
   "devDependencies": {
