+++ packages/logger/package.json
@@ -24,7 +24,7 @@
     "typescript": "~4.7.4"
   },
   "dependencies": {
-    "scramjet": "^4.36.9"
+    "scramjet": "^4.37.0"
   },
   "ava": {
     "extensions": [
