+++ packages/model/package.json
@@ -15,7 +15,7 @@
   "dependencies": {
     "@scramjet/obj-logger": "^0.37.0",
     "@scramjet/symbols": "^0.37.0",
-    "scramjet": "^4.36.9",
+    "scramjet": "^4.37.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
