+++ packages/runner/package.json
@@ -24,7 +24,7 @@
     "@scramjet/symbols": "^0.37.0",
     "@scramjet/utility": "^0.37.0",
     "bpmux": "^8.2.1",
-    "scramjet": "^4.36.9"
+    "scramjet": "^4.37.0"
   },
   "devDependencies": {
     "@scramjet/types": "^0.37.0",
