+++ packages/types/package.json
@@ -19,7 +19,7 @@
   },
   "devDependencies": {
     "@types/node": "15.12.5",
-    "scramjet": "^4.36.9",
+    "scramjet": "^4.37.0",
     "typedoc": "0.23.17",
     "typedoc-plugin-markdown": "3.13.6",
     "typescript": "~4.7.4"
